slightly less crazy
